refactor: 服薬記録コンポーネントの共通化
## 変更内容

### 新規作成
- `MedicationGroupedRecordsList` 共通コンポーネントを作成
  - PrescriptionBasedRecorder と RecordDetailView の重複ロジック・UIを統合
  - 柔軟な props でさまざまな用途に対応

### リファクタリング
- `PrescriptionBasedRecorder`: 264行 → 45行 (82%削減)
  - 共通コンポーネントを利用するラッパーに変更
- `RecordDetailView (DayRecordSection)`: 212行 → 28行 (87%削減)
  - 共通コンポーネントを利用するラッパーに変更

## 共通化された機能

### ロジック
- medications/records の取得
- medicationItems への展開
- 時間帯/処方箋でのグルーピング
- 記録状態の判定

### UI
- グループヘッダーとまとめて操作ボタン
- 薬リストのカード表示
- ローディング/空状態の表示

## Props による制御

- `allowGroupBySwitch`: グルーピング切り替えUI
- `showBulkActions`: まとめて操作表示
- `isEditable`: 編集可能制御
- `showUnrecordedStyle`: 未記録スタイル (solid/dashed)
- `showRecordDetails`: 服用時刻・メモ表示
- `showEditableBadge`: 編集可能バッジ表示

## メリット
✅ コード重複の大幅削減 (約450行削減)
✅ バグ修正が1箇所で済む
✅ 新機能の追加が容易
✅ DRY原則に従う
✅ 型安全性を維持

## テスト
- TypeScript 型チェック: ✅ 成功
- Biome フォーマット/リント: ✅ 成功
- Next.js ビルド: ✅ 成功

## 関連
- Closes #5
