+++ app/_shared/features/medication/components/index.ts
@@ -1,3 +1,4 @@
+export { MedicationGroupedRecordsList } from "./medication-grouped-records-list";
 export { MedicationRecordActions } from "./medication-record-actions";
 export { MedicationRecorder } from "./medication-recorder";
 export { PrescriptionBasedRecorder } from "./prescription-based-recorder";
